delete broblems in labRab2

+++ LabRab2/func.h
@@ -1,6 +1,6 @@
 #pragma once
 #include<fstream>
-int amtZero(int *, int);
-void sortColumn(int **, int, int);
-void initArr(int **, int , int, std::ifstream &);
-void printArr(int **, int, int, std::ofstream &);
+int amtZero(float *, int);
+void sortColumn(float **, int, int);
+void initArr(float **, int , int, std::ifstream &);
+void printArr(float **, int, int, std::ofstream &);
